server/pyproject.toml
[tool.poetry]
name = "Notes"
version = "0.0.1"
description = "Notes test task"
authors = ["Alexander Shcherbanyuk <@s0ul701>"]

[tool.poetry.dependencies]
python = "3.9.5"
Django = "3.0.5"
djongo = "^1.3.4"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.7.1"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.1"
django-extensions = "^3.1.3"
Werkzeug = "^2.0.1"
pytest-django = "^4.3.0"
model-bakery = "^1.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
